Sum of Selected Coefficients
INPUT & OUTPUT:
The input consists of multiple lines.
The first line contains a number n which indicates that the number of rows in the Pascal’s triangle will be n+1 (Note that rows of Pascal’s triangle are indexed starting with 0 at the top and the elements in a row are also indexed starting with a 0).
The second line contains a number m which indicates the number of transactions to be performed on the Pascal’s triangle. Each transaction is given in a separate line.  A transaction is a space separated list of integers. The first integer in each list indicates the row number, say R, and the rest of the integers in the list indicate the indices of values in row R. For each transaction, you have to compute the sum of given coefficients in the given row R.

Example: Input will be given in the following format:
5
3
3 1 2 
5 1 1 1 4
4 2 3 2