Polynomial Multiplication
This programming exercise is to give you exposure to the advantage of maintaining linked lists.  You should use the doubly linked list studied in the Week 3, first lecture.  The exercise is univariate polynomial multiplication.   Recall that a univariate polynomial is a polynomial over a single variable, and in the following discussion when we say polynomial, we mean a polynomial in a single variable.  You are given two polynomials – the first one, M1(x), is called the multiplicand, and the second one,  M2(x), is called the  multiplier. Your program has to output the product of M1(x) and M2(x).   You could use the functions implemented in the lectures to maintain your lists corresponding to the polynomials.  Below, you find a formal description of the inputs and outputs.

INPUT & OUTPUT:

The input will be four lines in the format specified below:

Line 1: Number of terms in multiplicand
e.g: 2

Line 2: Coefficients and exponents in the multiplicand in the format coeff 1 exp1 coeff2 exp2
e.g: 2 2 1 0 which represents 2x^2+1

Line 3: Number of terms in multiplier. 
e.g: 3

Line 4: Coefficients and exponents in the multiplier in the format coeff 1 exp1 coeff2 exp2 ....
e.g: -23 12 11 10 -6 41 which represents -6x^41-23x^12+11x^10

Each line will have multiple sets of coeff exp pairs, each corresponding to a non zero coefficient in the corresponding polynomial. The output has to be a single polynomial in ‘x’ that is the product polynomial. Note that, in the input, for each exponent, there may be multiple terms with the same exponent. However you should print a single term for each exponent while you print the output. Also, if the resulting product is zero, you have to explicitly print a zero.

The output has to be displayed as in the sample output shown with decreasing value of exponent.

Sample Input:
2
2 1 1 0
2
3 2 2 0

Sample Output:
6x^3+3x^2+4x+2